requirements.txt
Django==3.2.6
RPi.GPIO==0.7.0